New build update

+++ vk_api_closed.py
@@ -1,7 +1,5 @@
 # -*- coding: utf8 -*-
 
-#program updated and needed to upload github. new app id added. 01:23 19.12.2020
-
 import requests
 import sqlite3
 import json
@@ -52,7 +50,9 @@ massapps={}
 #id приложений
 try:
 	with open('appids.txt', 'r') as f:
-		app_ids = f.read().splitlines()
+		g=f.read().splitlines()
+		#print(g)
+		app_ids = g
 except Exception as e:
 	print(e)
 	app_ids=['7362610']
@@ -113,6 +113,17 @@ def set(status_id):
 	print("response: "+str(r.status_code)+" set id: "+str(status_id)+" Name of emoji: "+name)#+str(db['response']['status']['name']))
 
 	return json.loads(r.text)
+	
+def checkValidurl(url) -> bool:
+	try:
+		responce=requests.post(url)
+		responce=json.loads(responce.text)
+	
+		if(responce['error_description']=='application is disabled'):
+			return False
+	except Exception as e:
+		return True
+	return True
 
 
 #подключение БД
@@ -133,6 +144,7 @@ except Exception as e:
 choise=''
 update_tokens=''
 print_url=''
+info=False
 
 #меню
 settings=input("открыть отладочное меню?(y/n)")
@@ -141,6 +153,7 @@ if settings=='y':
 	print('2. вы хотите сбросить список таблицы?')
 	print('3. вы хотите сбросить персональные данные?(все сохраненные токены)')
 	print("4. Выводить ссылку эмодзи статуса")
+	print("5. вывести служебную информацию?(не рекоментуется к включению)")
 	print('0. выход')
 	choise=input()
 if choise=='1':
@@ -156,6 +169,9 @@ elif choise=='3':
 	except Exception as e:
 		print("personal info reset failed",e)
 elif choise=='4':
+	print_url='1'
+elif choise=='5':
+	info=True
 	print_url='1'
 try:
 	
@@ -177,46 +193,72 @@ for i in range(0, len(app_ids)):
 			basetimedatetime=datetime.strptime(basetime, "%d-%m-%Y %H:%M")
 
 		except Exception as e:
-			print("time don't installed, error text: ",e)
+			print("time don't installed for appid=",app_ids[i],"error text: ",e)
 			basetimedatetime=datetime.strptime("1-11-2077 1:27", "%d-%m-%Y %H:%M")
 		try:
 			baseip=str(c.execute(f'SELECT * FROM "personal" WHERE app_id="'+str(app_ids[i])+'" AND ip="'+ip+'";').fetchall()[0][3])
 		except Exception as e:
 			baseip="0.0.0.0"
-			print('Не получилось получить ip. сообщение ошибки: ',e)
+			print('Не получилось получить ip от приложения с appids=', app_ids[i],'сообщение ошибки: ',e)
 		try:
 			time_delta=time-basetimedatetime
 		except Exception as e:
 			print('opps, delta dont')
 			time_delta=time
-
-
+		for j in range (0,5):
+			if (time_delta.total_seconds() // 3600)>=24:
+				try:
+					c.execute('''DELETE FROM 'personal' WHERE app_id = ? AND ip=?''', (app_ids[i],ip))
+				except Exception as e:
+					print(app_ids[i],"time expired and falled him delete. e:",e)
+				time_delta= time
+		
+		if info:
+			print(baseip," ", ip)
+			print("update tokens var: ",update_tokens)
+			print("in personal table finded appids=",app_ids[i]," result:",is_findDB_exists('personal', 'app_id',app_ids[i]))
+			print("delta time >24?:",(time_delta.total_seconds() // 3600)>=24)
 
 		if ((not is_findDB_exists('personal', 'app_id',app_ids[i])) or (time_delta.total_seconds() // 3600)>=24 or update_tokens=='1' or not (ip==baseip)):
-			print('\n \n У вас истек или не обнаружен токен приложения: '+str(app_ids[i])+' требуется его получение, сейчас будет написана сыллка, вы должны раздрешить приложению доступ,потом когда будет написано, что нельзя никому её отправлять, скопируйте её сюда(это абсолютно безопасно, можете проверить исходник, этот файл). К сожалению токен живет 24 часа, по этому его переодически придется обновлять:( \n ')
-			print('https://oauth.vk.com/authorize?client_id='+str(app_ids[i])+'&scope=1024&redirect_uri=https://oauth.vk.com/blank.html&display=page&response_type=token&revoke=1 \n')
-			re=input('Вставте полученную ссылку: ')
-			token=re[re.find('access_token=')+13 : re.find('&')]
-			if not len(token)==85: 
-				print("Вы точно уверены, что ссылка содержит токен? Мы не смогли его вытащить")
-				token=input('Пожалуста, извлеките токен лично(скопируйте то, что между access_token= и &expires_in), вставте его сюда: ')
-				c.execute('''DELETE FROM 'personal' WHERE app_id = ? AND ip=?''', (app_ids[i],ip))
-			c.execute("""insert into 'personal' values (?,?,?,?)""",(
-			str(app_ids[i]),
-			token,
-			str(time.strftime("%d-%m-%Y %H:%M")),
-			ip
-			))
+			if checkValidurl('https://oauth.vk.com/authorize?client_id='+str(app_ids[i])+'&scope=1024&redirect_uri=https://oauth.vk.com/blank.html&display=page&response_type=token&revoke=1'):
+				print('\n \n У вас истек или не обнаружен токен приложения: '+str(app_ids[i])+' требуется его получение, сейчас будет написана сыллка, вы должны раздрешить приложению доступ,потом когда будет написано, что нельзя никому её отправлять, скопируйте её сюда(это абсолютно безопасно, можете проверить исходник, этот файл). К сожалению токен живет 24 часа, по этому его переодически придется обновлять:( \n ')
+				print('https://oauth.vk.com/authorize?client_id='+str(app_ids[i])+'&scope=1024&redirect_uri=https://oauth.vk.com/blank.html&display=page&response_type=token&revoke=1 \n')
+				re=input('Вставте полученную ссылку: ')
+				if re=='' or re=='-1':
+					print("\n пропускаю токен \n")
+				else:	
+					token=re[re.find('access_token=')+13 : re.find('&')]
+					if not len(token)==85: 
+						print("Вы точно уверены, что ссылка содержит токен? Мы не смогли его вытащить")
+						token=input('Пожалуста, извлеките токен лично(скопируйте то, что между access_token= и &expires_in), вставте его сюда: ')
+						c.execute('''DELETE FROM 'personal' WHERE app_id = ? AND ip=?''', (app_ids[i],ip))
+					c.execute("""insert into 'personal' values (?,?,?,?)""",(
+					str(app_ids[i]),
+					token,
+					str(time.strftime("%d-%m-%Y %H:%M")),
+					ip
+					))
 
-conn.commit()
+					conn.commit()
+			else:
+				print("\n appid",app_ids[i],'is not valid \n')
+
 
 for i in range(0, len(app_ids)):
-	massapps[str(app_ids[i])]['access_token']=str(c.execute(f'SELECT * FROM "personal" WHERE app_id="'+str(app_ids[i])+'" AND ip="'+ip+'";').fetchall()[0][1])
+	try:
+	   massapps[str(app_ids[i])]['access_token']=str(c.execute(f'SELECT * FROM "personal" WHERE app_id="'+str(app_ids[i])+'" AND ip="'+ip+'";').fetchall()[0][1])
+	except Exception as e:
+	   print('access token не получен из базы, от приложения',app_ids[i],'\n')
+	
 
 
 
-if input("Мне проверить новые эмодзи?(если запуск впервые, то соглашайтесь) (y/n)")=="y":
+if input("Мне проверить новые эмодзи?(если запуск впервые, то соглашайтесь) (y/n) ")=="y":
 	print("проверка на наличие новых эмодзи")
+	try:
+		c.execute('''DROP TABLE IF EXISTS 'Base';''')
+	except Exception as e:
+		print("table reset failed",e)
 	for i in range(0, len(app_ids)):
 		db=post(app_ids[i])
 		addtoDB(db,app_ids[i])
